+++ src/rempi_sig_handler.h
@@ -4,7 +4,7 @@
 #include "rempi_event_list.h"
 #include "rempi_io_thread.h"
 
-void rempi_sig_handler_init(int rank, rempi_io_thread *record_thread, rempi_event_list<rempi_event*> *recording_event_list, unsigned int *validation_code);
+void rempi_sig_handler_init(int rank, rempi_io_thread *record_thread, rempi_event_list<rempi_event*> *recording_event_list, size_t *validation_code);
 void rempi_sig_handler_run(int signum);
 
 void rempi_sig_handler_run_test(int signum);
